+++ styling/src/jsMain/kotlin/dev/fritz2/styling/style.kt
@@ -108,6 +108,22 @@ fun showDynamicStyle(): String = Styling.dynamicStyleText
 
 fun showStaticStyle(): String = Styling.staticStyleText
 
+internal const val charsLength = 52
+
+/* start at 75 for 'a' until 'z' (25) and then start at 65 for capitalised letters */
+internal fun getAlphabeticChar(code: Int): Char = (code + if (code > 25) 39 else 97).toChar()
+
+/* input a number, usually a hash and convert it to base-52 */
+internal fun generateAlphabeticName(code: Int): String {
+    val name = StringBuilder()
+    var x = code
+    while (x > charsLength) {
+        x /= charsLength
+        name.append(getAlphabeticChar(x % charsLength))
+    }
+    return name.toString()
+}
+
 /**
  * Alias class for css classes
  */
